Fix for overloading problems

+++ src/costmap_2d/include/costmap_2d/costmap_2d.h
@@ -372,7 +372,7 @@ protected:
       unsigned int offset = y0 * size_x_ + x0;
 
       // we need to chose how much to scale our dominant dimension, based on the maximum length of the line
-      double dist = hypot(dx, dy);
+      double dist = hypot((double)dx, (double)dy);
       double scale = (dist == 0.0) ? 1.0 : std::min(1.0, max_length / dist);
 
       // if x is dominant
